misc: addressed review comments
